+++ packages/tailwindcss/src/index.test.ts
@@ -5965,4 +5965,18 @@ describe('feature detection', () => {
     expect(compiler.features & Features.AtImport).toBeTruthy()
     expect(compiler.features & Features.Utilities).toBeFalsy()
   })
+
+  test('using `@theme`', async () => {
+    let compiler = await compile(css`
+      @theme {
+        --tracking-narrower: -0.02em;
+      }
+    `)
+
+    // We see @theme
+    expect(compiler.features & Features.AtTheme).toBeTruthy()
+
+    // And this produces no output because no other CSS is present
+    expect(compiler.build([])).toEqual('')
+  })
 })